api-auth/pyproject.toml
[tool.poetry]
name = "api-auth"
version = "0.1.0"
description = ""
authors = ["Davi Lucciola <davilucciola14@gmail.com>"]
readme = "../README.md"
packages = [{include = "api"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.2"
uvicorn = "^0.22.0"
sqlmodel = "^0.0.8"
python-jose = "^3.3.0"
passlib = "^1.7.4"
alembic = "^1.11.1"
pydantic = {extras = ["email"], version = "^1.10.7"}
python-decouple = "^3.8"
psycopg2-binary = "^2.9.6"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
